fix: unify admin wallet check so feature flags page works for all admin address formats
The feature flags API used simple exact-match for admin wallets while
the admin check API derived stake addresses from payment addresses.
Extract shared isAdminWallet into lib/adminAuth.ts so both endpoints
recognize admins consistently.
